:wrench: chore: 프론트 - 강사님 코드 구조로 수정
